requirements.txt
psycopg2-binary==2.9.11
PySide6==6.10.2
PySide6_Addons==6.10.2
PySide6_Essentials==6.10.2
shiboken6==6.10.2
